requirements.txt
sentence_transformers
transformers
pysqlite3
openai
rank_bm25
spacy
nltk
kss
jsonlines
pandas
openpyxl
google-generativeai
sentencepiece

# python>=3.9
# sentence_transformers>=2.2.2
# transformers>=4.29.2
# torch>=>=2.0.0
# google-generativeai>=0.5.0